requirements.txt
requests~=2.32
docker~=7.1
pyyaml~=6.0